+++ src/importexport/midi/CMakeLists.txt
@@ -20,10 +20,12 @@
 
 set(MODULE iex_midi)
 
+include(${CMAKE_CURRENT_LIST_DIR}/internal/midishared/midishared.cmake)
 include(${CMAKE_CURRENT_LIST_DIR}/internal/midiimport/midiimport.cmake)
 include(${CMAKE_CURRENT_LIST_DIR}/internal/midiexport/midiexport.cmake)
 
 set(MODULE_SRC
+    ${MIDISHARED_SRC}
     ${MIDIIMPORT_SRC}
     ${MIDIEXPORT_SRC}
     ${CMAKE_CURRENT_LIST_DIR}/midimodule.cpp
@@ -38,7 +40,7 @@ set(MODULE_SRC
     )
 
 set(MODULE_LINK
-    beatroot    # for midi
+    beatroot
     engraving
     qzip
     )
