+++ src/importexport/midi/internal/midiimport/importmidi_chordname.cpp
@@ -24,12 +24,13 @@
 #include "importmidi_chord.h"
 #include "importmidi_fraction.h"
 #include "importmidi_operations.h"
+#include "../midishared/midifile.h"
+
 #include "engraving/libmscore/masterscore.h"
 #include "engraving/libmscore/staff.h"
 #include "engraving/libmscore/measure.h"
 #include "engraving/libmscore/harmony.h"
 #include "engraving/libmscore/factory.h"
-#include "engraving/compat/midi/midifile.h"
 
 // From XF Format Specifications V 2.01 (January 13, 1999, YAMAHA CORPORATION)
