+++ src/importexport/midi/internal/midishared/midishared.cmake
@@ -1,0 +1,4 @@
+set (MIDISHARED_SRC
+    ${CMAKE_CURRENT_LIST_DIR}/midifile.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/midifile.h
+    )
